+++ Alba/Controllers/ClientController.cs
@@ -44,7 +44,7 @@ namespace TerritoryTools.Alba.Controllers
                     return;
                 }
 
-                new AddressImporter(AuthenticatedClient(), int.Parse(view.UploadDelayMs), languageFilePath)
+                new AddressImporter(AuthenticatedClient(), int.Parse(view.UploadDelayMs), languageFilePath: languageFilePath)
                     .AddFrom(addressFilePath);
             }
             catch (UserException e)