+++ Alba/PowerShell/AlbaRemote/EditAlbaAddress.cs
@@ -1,21 +1,13 @@
 ﻿using Controllers.AlbaServer;
 using System;
-using System.Collections.Generic;
 using System.Management.Automation;
-using TerritoryTools.Alba.Controllers.AlbaServer;
 using TerritoryTools.Alba.Controllers.UseCases;
 
 namespace TerritoryTools.Alba.PowerShell
 {
     [Cmdlet(VerbsData.Edit,"AlbaAddress")]
-    public class EditAlbaAddress : PSCmdlet
+    public class EditAlbaAddress : AlbaConnectedCmdlet
     {
-        [Parameter]
-        public string LanguageFilePath { get; set; }
-
-        [Parameter]
-        public AlbaConnection Connection { get; set; }
-
         [Parameter(
             Mandatory = false,
             Position = 0,
@@ -30,19 +22,10 @@ namespace TerritoryTools.Alba.PowerShell
 
         protected override void BeginProcessing()
         {
-            if (Connection == null)
-            {
-                Connection = SessionState
-                    .PSVariable
-                    .Get(nameof(Names.CurrentAlbaConnection))?
-                    .Value as AlbaConnection
-                    ?? throw new MissingConnectionException();
-            }
-
             importer = new AddressImporter(
                 Connection, 
-                UploadDelayMs, 
-                LanguageFilePath);
+                UploadDelayMs,
+                languages: Languages);
         }
 
         protected override void ProcessRecord()