+++ Alba/PowerShell/AlbaRemote/GetAlbaTerritory.cs
@@ -7,24 +7,12 @@ namespace TerritoryTools.Alba.PowerShell
 {
     [Cmdlet(VerbsCommon.Get, "AlbaTerritory")]
     [OutputType(typeof(Assignment))]
-    public class GetAlbaTerritory : PSCmdlet
+    public class GetAlbaTerritory : AlbaConnectedCmdlet
     {
-        [Parameter]
-        public AlbaConnection Connection { get; set; }
-
         protected override void ProcessRecord()
         {
             try
             {
-                if (Connection == null)
-                {
-                    Connection = SessionState
-                        .PSVariable
-                        .Get(nameof(Names.CurrentAlbaConnection))?
-                        .Value as AlbaConnection
-                        ?? throw new MissingConnectionException();
-                }
-
                 var assignmentsResultString = Connection.DownloadString(
                    RelativeUrlBuilder.GetTerritoryAssignments());