+++ Alba/PowerShell/AlbaRemote/MissingLanguagesException.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace TerritoryTools.Alba.PowerShell
+{
+    public class MissingLanguagesException : Exception
+    {
+        public override string Message =>
+            "Run Get-AlbaLanguage to load a list of language IDs from Alba";
+    }
+}