+++ Alba/PowerShell/ExampleScripts/Download-Addresses.ps1
@@ -1,9 +1,7 @@
 ﻿
 $connection = Get-AlbaConnection `
     -AlbaHost $env:ALBA_HOST `
-    -Account $env:ALBA_ACCOUNT `
-    -User $env:ALBA_USER `
-    -Password $env:ALBA_PASSWORD
+    -Account $env:ALBA_ACCOUNT
 
 $result = Get-AlbaAddress -Connection $connection -Search "Main St"