+++ Alba/PowerShell/PublishTo-PSGallery.ps1
@@ -1,6 +1,11 @@
 ﻿$folderPath = "$PSScriptRoot\publish\TerritoryTools"
 $timestamp = (Get-Date -Format "yyyy-MM-dd-HHmmss")
 
+If(!$env:PSGALLERY_NUGETAPIKEY) {
+    throw "Could not read NuGet API Key for PowerShell Gallery environment variable named PSGALLERY_NUGETAPIKEY"
+    exit
+}
+
 Publish-Module `
   -Path $folderPath  `
   -NuGetApiKey $env:PSGALLERY_NUGETAPIKEY `