+++ Alba/PowerShell/HISTORY.md
@@ -1,4 +1,15 @@
 # Change Log
+## 0.1.15
+### Features
+- Renamed Scripts folder to ExampleScripts to describe what the scripts are for
+- Add `Get-AlbaLanguage` Cmdlet
+- Add `Get-AlbaLangaugeId` Cmdlet
+- Add `Clear-AlbaTerritoryUser` Cmdlet
+
+## 0.1.0 to 0.1.14
+### Features
+- Add `PublishTo-PSGallery.ps1` script
+
 ## 0.1.0
 ### Features
 - This is a first attempt at creating a PowerShell module for Territory Tools